fixed as per comments. units tests are added as seperate package
